Move test and add another
