+++ Tests/CSharp/ExpressionTests.cs
@@ -2340,6 +2340,94 @@ End Class", @"public partial class Class1
         }
 
         [Fact]
+        public async Task TestGenericMethodGroupGainsBrackets()
+        {
+            //BUG: Comment after New With is lost
+            await TestConversionVisualBasicToCSharpWithoutComments(
+                @"Public Enum TheType
+    Tree
+End Enum
+
+Public Class MoreParsing
+    Sub DoGet()
+        Dim anon = New With {
+            .TheType = GetEnumValues(Of TheType)
+        }
+    End Sub
+
+    Private Function GetEnumValues(Of TEnum)() As IDictionary(Of Integer, String)
+        Return System.Enum.GetValues(GetType(TEnum)).Cast(Of TEnum).
+            ToDictionary(Function(enumValue) DirectCast(DirectCast(enumValue, Object), Integer),
+                         Function(enumValue) enumValue.ToString())
+    End Function
+End Class",
+                @"using System;
+using System.Collections.Generic;
+using System.Linq;
+
+public enum TheType
+{
+    Tree
+}
+
+public partial class MoreParsing
+{
+    public void DoGet()
+    {
+        var anon = new
+        {
+            TheType = MoreParsing.GetEnumValues<TheType>()
+        };
+    }
+
+    private IDictionary<int, string> GetEnumValues<TEnum>()
+    {
+        return Enum.GetValues(typeof(TEnum)).Cast<TEnum>().ToDictionary(enumValue => (int)(object)enumValue, enumValue => enumValue.ToString());
+    }
+}");
+        }
+
+        [Fact]
+        public async Task GenericMethodCalledWithAnonymousType()
+        {
+            //BUG: Comment after New With is lost
+            await TestConversionVisualBasicToCSharpWithoutComments(
+                @"Public Class MoreParsing
+    Sub DoGet()
+        Dim anon = New With {
+            .ANumber = 5
+        }
+        Dim sameAnon = Identity(anon)
+        Dim repeated = Enumerable.Repeat(anon, 5).ToList()
+    End Sub
+
+    Private Function Identity(Of TType)(tInstance As TType) As TType
+        Return tInstance
+    End Function
+End Class",
+                @"using System.Linq;
+using Microsoft.VisualBasic.CompilerServices;
+
+public partial class MoreParsing
+{
+    public void DoGet()
+    {
+        var anon = new
+        {
+            ANumber = 5
+        };
+        var sameAnon = this.Identity(anon);
+        var repeated = Enumerable.Repeat(anon, Conversions.ToInteger(5)).ToList();
+    }
+
+    private TType Identity<TType>(TType tInstance)
+    {
+        return tInstance;
+    }
+}");
+        }
+
+        [Fact]
         public async Task AliasedImportsWithTypePromotionIssue401()
         {
             await TestConversionVisualBasicToCSharpWithoutComments(
